refactor: Move UI's opening and group packages up a step.
