+++ app/src/main/java/no/usn/mob3000/ui/components/base/Routes.kt
@@ -28,13 +28,13 @@ import no.usn.mob3000.ui.screens.auth.LoginScreen
 import no.usn.mob3000.ui.screens.auth.ResetPasswordScreen
 import no.usn.mob3000.ui.screens.chess.HistoryScreen
 import no.usn.mob3000.ui.screens.chess.PlayScreen
-import no.usn.mob3000.ui.screens.chess.train.group.CreateGroupScreen
-import no.usn.mob3000.ui.screens.chess.train.group.GroupsScreen
-import no.usn.mob3000.ui.screens.chess.train.group.UpdateGroupScreen
-import no.usn.mob3000.ui.screens.chess.train.opening.CreateOpeningScreen
-import no.usn.mob3000.ui.screens.chess.train.opening.OpeningDetailsScreen
-import no.usn.mob3000.ui.screens.chess.train.opening.OpeningsScreen
-import no.usn.mob3000.ui.screens.chess.train.opening.UpdateOpeningScreen
+import no.usn.mob3000.ui.screens.chess.group.CreateGroupScreen
+import no.usn.mob3000.ui.screens.chess.group.GroupsScreen
+import no.usn.mob3000.ui.screens.chess.group.UpdateGroupScreen
+import no.usn.mob3000.ui.screens.chess.opening.CreateOpeningScreen
+import no.usn.mob3000.ui.screens.chess.opening.OpeningDetailsScreen
+import no.usn.mob3000.ui.screens.chess.opening.OpeningsScreen
+import no.usn.mob3000.ui.screens.chess.opening.UpdateOpeningScreen
 import no.usn.mob3000.ui.screens.info.AboutUsScreen
 import no.usn.mob3000.ui.screens.info.InfoScreen
 import no.usn.mob3000.ui.screens.info.docs.CreateDocumentationScreen
